Move occupancy related functions into MemUtils
